feat[refactor]: clean up unused components, consolidate utils, and remove stale assets
- Remove unused components (CardSwap, FlowingMenu, ImpactCardSwap, LoadingScreen, ProductShowcase, TeamSection, KTEDContact)
- Delete stale plan docs and product images
- Move utils: lib/utils.ts → utils/cn.ts, emailjs.js → emailjs.ts
- Consolidate CSS: globals.css → global.css
- Rename privacy policy content files to drop language suffix
- Remove package-lock.json (using pnpm)
- Update components.json aliases and CSS paths
- Minor UI tweaks across Footer, Hero, ServicesShowcase, Contact sections
