--- docs/services.md
@@ -1,50 +1,0 @@
-# Servicios
-## Desarrollo a Medida
-Diseñamos y desarrollamos software personalizado con inteligencia artificial para resolver problemas concretos de tu negocio. Desde aplicaciones móviles y plataformas web hasta agentes inteligentes u otras automatizaciones. No vendemos soluciones genéricas: analizamos tu operativa y construimos exactamente lo que necesitas.
-
-## Conexion Kit Espacio de Datos Europeo
-Te ayudamos a incorporar tu empresa a los espacios de datos sectoriales europeos (agroalimentario, industria, salud, comercio) y a acceder a las ayudas públicas del programa KTED, con hasta 30.000€ de financiación. Nos encargamos de todo lo técnico: preparación de datos, desarrollo del conector, documentación y registro en el espacio de datos.
-
-## Consultoria de Transformacion Digital 
-Visitamos tu empresa, entendemos cómo trabajáis de verdad y os decimos con honestidad qué se puede mejorar con tecnología y qué no. En menos de una semana recibes un informe con oportunidades concretas, estimación de ahorro y un plan de acción priorizado. Si no necesitas nada, te lo decimos. Sin compromiso, sin humo. 
-
-## Consultoria de Innovacion
-Identificamos las ayudas públicas que encajan con tu empresa y nos encargamos de diseñar el proyecto para que cumpla todos los requisitos de la convocatoria. Preparamos la memoria técnica, el presupuesto y toda la documentación necesaria para maximizar tus posibilidades de aprobación. IGAPE, IA360, Ticket Innova, Kit Digital, espacios de datos... conocemos las convocatorias porque trabajamos con ellas a diario.
-
-# Proyectos
-## TAKE 
-Implementación de nuestra plataforma BOND en la cafetería Take. Tarjeta de fidelización digital con Google y Apple Wallet, sistema de puntos automático y dashboard de analíticas de consumo. Primer piloto real del producto en hostelería, validando el modelo con clientes reales.
-
-## Consultoría especializada
-Implantación de Nessie Oficina para una consultora especializada en metodologías relacionales. Gestión documental con IA, integración con email, búsqueda semántica sobre toda la documentación de la empresa y sistema de unificación de carteras de clientes con matching de subvenciones.
-
-## ICIGA
-Implementación de nuestra plataforma IRIS para un distribuidor comercial de material de laboratorio. Monitoriza 6 bandejas de entrada, clasifica automáticamente solicitudes de presupuesto, pedidos y facturas mediante IA, y cruza cada petición contra una base de datos de más de 63.000 productos químicos. Genera propuestas de respuesta listas para revisión humana e integra con el ERP de la empresa.
-
-## Shearn
-Plataforma propia para la resolución de ejercicios de dibujo técnico. Motor 2D que genera soluciones paso a paso y módulo 3D en Three.js para la reconstrucción de sólidos a partir de vistas ortográficas. Desarrollado íntegramente de forma interna como herramienta educativa.
-
-## Plataforma de gamificación para academias de oposiciones
-Sistema completo de gamificación para una plataforma de oposiciones online: red social con amigos y retos 1v1, ligas competitivas con temporadas y sistema ELO, economía de puntos y monedas, coach virtual con rachas motivacionales, catálogo de premios canjeables y panel de administración. Diseño mobile-first con Material Design, integrado con la infraestructura existente del cliente en Laravel y Odoo.
-
-## Grille Sociedade Cooperativa Galega
-Aplicación móvil de evaluación del rendimiento para el sector agrícola. Permite a supervisores puntuar a trabajadores con criterios personalizables y pesos configurables, generando informes PDF consolidados por periodo. Diseñada para uso en campo con interfaz de alto contraste, controles táctiles amplios y funcionamiento offline, disponible en castellano y gallego.
-
-# Productos
-## APEC - Agent-Powered Enterprise Communications
-Automatiza tu prospección comercial sin perder el toque humano. APEC investiga empresas, extrae datos financieros, evalúa el encaje con tu oferta mediante IA y genera emails de contacto hiperpersonalizados. Su dashboard centraliza envíos, seguimiento de aperturas y clics, y programación de follow-ups. Escala tu captación B2B de horas de trabajo manual a minutos de supervisión.
-
-## Nessie
-Conecta toda la documentación de tu empresa y pregúntale lo que quieras en lenguaje natural. Nessie procesa automáticamente PDFs, Word, hojas de cálculo y más desde Google Drive, los convierte en conocimiento buscable con IA y te responde en tiempo real con referencias a los documentos originales. Disponible en versión Oficina para consultoras y despachos, e Industrial para plantas de producción con consulta por voz y acceso desde tablet.
-
-## Moura
-Encuentra las subvenciones que encajan con tu empresa sin buscar. Moura cruza automáticamente las convocatorias públicas con tu perfil empresarial — sector, ubicación, tamaño — y analiza las bases con IA para extraer requisitos, plazos y presupuestos. Cuando hay un match, genera la memoria técnica, el plan de trabajo y el presupuesto detallado con un solo clic.
-
-## ACTA — Automated Capture, Transcription & Analysis
-Graba tus reuniones, formaciones o entrevistas y obtén transcripciones profesionales con identificación automática de quién habla en cada momento. Echoia combina transcripción de alta precisión con diarización inteligente y agentes de resumen que extraen las decisiones clave, tareas pendientes y puntos de acción de cada conversación.
-
-## IRIS — Intelligent Response & Inbox System
-Un sistema de IA que monitoriza tus bandejas de entrada, clasifica automáticamente cada email por tipo (presupuestos, pedidos, facturas, consultas), lo cruza contra tus bases de datos internas y genera propuestas de respuesta listas para tu revisión. Se integra con tu ERP y CRM. Transforma horas de gestión de correo en minutos de supervisión.
-
-## BOND - Business-Owned Network for Digital loyalty
-Una app de fidelización bajo tu propia marca con tarjeta de puntos digital integrada en Google y Apple Wallet. Tus clientes acumulan puntos automáticamente, reciben ofertas personalizadas y tú accedes a un dashboard con analíticas de consumo en tiempo real. Sin tarjetas de papel, sin comisiones de plataformas externas, sin dependencia de terceros.
