+++ public/sitemap.xml
@@ -64,10 +64,6 @@
     <priority>0.8</priority>
     <xhtml:link rel="alternate" hreflang="es" href="https://osix.tech/es/projects/nessie"/>
     <xhtml:link rel="alternate" hreflang="en" href="https://osix.tech/en/projects/nessie"/>
-    <image:image>
-      <image:loc>https://osix.tech/projects/nessie.png</image:loc>
-      <image:title>Nessie Assistant - AI Document Assistant</image:title>
-    </image:image>
   </url>
 
   <!-- Project: Nessie (Spanish) -->
@@ -78,10 +74,6 @@
     <priority>0.8</priority>
     <xhtml:link rel="alternate" hreflang="en" href="https://osix.tech/en/projects/nessie"/>
     <xhtml:link rel="alternate" hreflang="es" href="https://osix.tech/es/projects/nessie"/>
-    <image:image>
-      <image:loc>https://osix.tech/projects/nessie.png</image:loc>
-      <image:title>Nessie Assistant - Asistente IA de Documentos</image:title>
-    </image:image>
   </url>
 
   <!-- Project: TAKE (English) -->
@@ -93,7 +85,7 @@
     <xhtml:link rel="alternate" hreflang="es" href="https://osix.tech/es/projects/take"/>
     <xhtml:link rel="alternate" hreflang="en" href="https://osix.tech/en/projects/take"/>
     <image:image>
-      <image:loc>https://osix.tech/projects/take.jpg</image:loc>
+      <image:loc>https://osix.tech/projects/take.png</image:loc>
       <image:title>TAKE - AI Ticket Management System</image:title>
     </image:image>
   </url>
@@ -107,7 +99,7 @@
     <xhtml:link rel="alternate" hreflang="en" href="https://osix.tech/en/projects/take"/>
     <xhtml:link rel="alternate" hreflang="es" href="https://osix.tech/es/projects/take"/>
     <image:image>
-      <image:loc>https://osix.tech/projects/take.jpg</image:loc>
+      <image:loc>https://osix.tech/projects/take.png</image:loc>
       <image:title>TAKE - Sistema IA de Gestión de Tickets</image:title>
     </image:image>
   </url>
