--- src/components/CardSwap.jsx
@@ -1,208 +1,0 @@
-import React, { Children, cloneElement, forwardRef, isValidElement, useEffect, useImperativeHandle, useMemo, useRef } from 'react';
-import gsap from 'gsap';
-
-export const Card = forwardRef(({ customClass, ...rest }, ref) => (
-  <div
-    ref={ref}
-    {...rest}
-    className={`card-swap-card ${customClass ?? ''} ${rest.className ?? ''}`.trim()}
-  />
-));
-Card.displayName = 'Card';
-
-const makeSlot = (i, distX, distY, total) => ({
-  x: i * distX,
-  y: -i * distY,
-  z: -i * distX * 1.5,
-  zIndex: total - i,
-});
-
-const placeNow = (el, slot, skew) =>
-  gsap.set(el, {
-    x: slot.x,
-    y: slot.y,
-    z: slot.z,
-    xPercent: -50,
-    yPercent: -50,
-    skewY: skew,
-    transformOrigin: 'center center',
-    zIndex: slot.zIndex,
-    force3D: true,
-  });
-
-const CONFIGS = {
-  elastic: {
-    ease: 'elastic.out(0.6,0.9)',
-    durDrop: 2,
-    durMove: 2,
-    durReturn: 2,
-    promoteOverlap: 0.9,
-    returnDelay: 0.05,
-  },
-  smooth: {
-    ease: 'power1.inOut',
-    durDrop: 0.8,
-    durMove: 0.8,
-    durReturn: 0.8,
-    promoteOverlap: 0.45,
-    returnDelay: 0.2,
-  },
-};
-
-const CardSwap = forwardRef(({
-  width = 500,
-  height = 400,
-  cardDistance = 60,
-  verticalDistance = 70,
-  delay = 5000,
-  pauseOnHover = false,
-  swapOnClick = false,
-  onCardClick,
-  skewAmount = 6,
-  easing = 'elastic',
-  children,
-}, fwdRef) => {
-  const config = CONFIGS[easing] || CONFIGS.elastic;
-
-  const childArr = useMemo(() => Children.toArray(children), [children]);
-  const refs = useMemo(
-    () => childArr.map(() => React.createRef()),
-    [childArr.length]
-  );
-
-  const order = useRef(Array.from({ length: childArr.length }, (_, i) => i));
-  const tlRef = useRef(null);
-  const intervalRef = useRef();
-  const container = useRef(null);
-
-  const runSwap = () => {
-    if (order.current.length < 2) return;
-
-    const [front, ...rest] = order.current;
-    const elFront = refs[front].current;
-    const tl = gsap.timeline();
-    tlRef.current = tl;
-
-    tl.to(elFront, {
-      y: '+=500',
-      duration: config.durDrop,
-      ease: config.ease,
-    });
-
-    tl.addLabel('promote', `-=${config.durDrop * config.promoteOverlap}`);
-    rest.forEach((idx, i) => {
-      const el = refs[idx].current;
-      const slot = makeSlot(i, cardDistance, verticalDistance, refs.length);
-      tl.set(el, { zIndex: slot.zIndex }, 'promote');
-      tl.to(
-        el,
-        { x: slot.x, y: slot.y, z: slot.z, duration: config.durMove, ease: config.ease },
-        `promote+=${i * 0.15}`
-      );
-    });
-
-    const backSlot = makeSlot(refs.length - 1, cardDistance, verticalDistance, refs.length);
-    tl.addLabel('return', `promote+=${config.durMove * config.returnDelay}`);
-    tl.call(() => gsap.set(elFront, { zIndex: backSlot.zIndex }), undefined, 'return');
-    tl.to(
-      elFront,
-      { x: backSlot.x, y: backSlot.y, z: backSlot.z, duration: config.durReturn, ease: config.ease },
-      'return'
-    );
-
-    tl.call(() => {
-      order.current = [...rest, front];
-    });
-  };
-
-  const forceFinishAndSwap = () => {
-    if (tlRef.current) {
-      tlRef.current.progress(1, false);
-      tlRef.current.kill();
-      tlRef.current = null;
-    }
-    runSwap();
-  };
-
-  useImperativeHandle(fwdRef, () => ({
-    swap: () => {
-      clearInterval(intervalRef.current);
-      forceFinishAndSwap();
-      if (delay > 0) {
-        intervalRef.current = window.setInterval(() => runSwap(), delay);
-      }
-    },
-  }));
-
-  useEffect(() => {
-    const total = refs.length;
-    refs.forEach((r, i) =>
-      placeNow(r.current, makeSlot(i, cardDistance, verticalDistance, total), skewAmount)
-    );
-
-    if (delay > 0) {
-      runSwap();
-      intervalRef.current = window.setInterval(() => runSwap(), delay);
-    }
-
-    if (pauseOnHover) {
-      const node = container.current;
-      const pause = () => {
-        tlRef.current?.pause();
-        clearInterval(intervalRef.current);
-      };
-      const resume = () => {
-        tlRef.current?.play();
-        if (delay > 0) {
-          intervalRef.current = window.setInterval(() => runSwap(), delay);
-        }
-      };
-      node.addEventListener('mouseenter', pause);
-      node.addEventListener('mouseleave', resume);
-      return () => {
-        node.removeEventListener('mouseenter', pause);
-        node.removeEventListener('mouseleave', resume);
-        clearInterval(intervalRef.current);
-      };
-    }
-    return () => clearInterval(intervalRef.current);
-  }, []);
-
-  const handleClick = (i) => {
-    if (!swapOnClick) return;
-    clearInterval(intervalRef.current);
-    forceFinishAndSwap();
-    if (delay > 0) {
-      intervalRef.current = window.setInterval(() => runSwap(), delay);
-    }
-  };
-
-  const rendered = childArr.map((child, i) =>
-    isValidElement(child)
-      ? cloneElement(child, {
-          key: i,
-          ref: refs[i],
-          style: {
-            width,
-            height,
-            cursor: swapOnClick ? 'pointer' : 'default',
-            ...(child.props.style ?? {}),
-          },
-          onClick: (e) => {
-            child.props.onClick?.(e);
-            onCardClick?.(i);
-            handleClick(i);
-          },
-        })
-      : child
-  );
-
-  return (
-    <div ref={container} className="card-swap-container" style={{ width, height }}>
-      {rendered}
-    </div>
-  );
-});
-CardSwap.displayName = 'CardSwap';
-
-export default CardSwap;
