--- src/components/FlowingMenu.jsx
@@ -1,161 +1,0 @@
-import { useRef, useEffect, useState } from 'react';
-import { gsap } from 'gsap';
-
-function FlowingMenu({
-  items = [],
-  speed = 15,
-  textColor = '#000',
-  bgColor = '#fff',
-  marqueeBgColor = '#000',
-  marqueeTextColor = '#fff',
-  borderColor = '#e5e7eb'
-}) {
-  return (
-    <div className="flowing-menu-wrap" style={{ backgroundColor: bgColor }}>
-      <nav className="flowing-menu">
-        {items.map((item, idx) => (
-          <MenuItem
-            key={idx}
-            {...item}
-            index={idx}
-            speed={speed}
-            textColor={textColor}
-            marqueeBgColor={marqueeBgColor}
-            marqueeTextColor={marqueeTextColor}
-            borderColor={borderColor}
-          />
-        ))}
-      </nav>
-    </div>
-  );
-}
-
-function MenuItem({ link, text, subtitle, image, index, speed, textColor, marqueeBgColor, marqueeTextColor, borderColor }) {
-  const isEven = index % 2 === 1;
-  const itemRef = useRef(null);
-  const marqueeRef = useRef(null);
-  const marqueeInnerRef = useRef(null);
-  const animationRef = useRef(null);
-  const [repetitions, setRepetitions] = useState(4);
-
-  const animationDefaults = { duration: 0.6, ease: 'expo' };
-
-  const findClosestEdge = (mouseX, mouseY, width, height) => {
-    const topEdgeDist = distMetric(mouseX, mouseY, width / 2, 0);
-    const bottomEdgeDist = distMetric(mouseX, mouseY, width / 2, height);
-    return topEdgeDist < bottomEdgeDist ? 'top' : 'bottom';
-  };
-
-  const distMetric = (x, y, x2, y2) => {
-    const xDiff = x - x2;
-    const yDiff = y - y2;
-    return xDiff * xDiff + yDiff * yDiff;
-  };
-
-  useEffect(() => {
-    const calculateRepetitions = () => {
-      if (!marqueeInnerRef.current) return;
-      const marqueeContent = marqueeInnerRef.current.querySelector('.flowing-marquee__part');
-      if (!marqueeContent) return;
-      const contentWidth = marqueeContent.offsetWidth;
-      const viewportWidth = window.innerWidth;
-      const needed = Math.ceil(viewportWidth / contentWidth) + 2;
-      setRepetitions(Math.max(4, needed));
-    };
-
-    calculateRepetitions();
-    window.addEventListener('resize', calculateRepetitions);
-    return () => window.removeEventListener('resize', calculateRepetitions);
-  }, [text, image]);
-
-  useEffect(() => {
-    const setupMarquee = () => {
-      if (!marqueeInnerRef.current) return;
-      const marqueeContent = marqueeInnerRef.current.querySelector('.flowing-marquee__part');
-      if (!marqueeContent) return;
-      const contentWidth = marqueeContent.offsetWidth;
-      if (contentWidth === 0) return;
-
-      if (animationRef.current) animationRef.current.kill();
-
-      animationRef.current = gsap.to(marqueeInnerRef.current, {
-        x: -contentWidth,
-        duration: speed,
-        ease: 'none',
-        repeat: -1
-      });
-    };
-
-    const timer = setTimeout(setupMarquee, 50);
-    return () => {
-      clearTimeout(timer);
-      if (animationRef.current) animationRef.current.kill();
-    };
-  }, [text, image, repetitions, speed]);
-
-  const handleMouseEnter = (ev) => {
-    if (!itemRef.current || !marqueeRef.current || !marqueeInnerRef.current) return;
-    const rect = itemRef.current.getBoundingClientRect();
-    const x = ev.clientX - rect.left;
-    const y = ev.clientY - rect.top;
-    const edge = findClosestEdge(x, y, rect.width, rect.height);
-
-    gsap
-      .timeline({ defaults: animationDefaults })
-      .set(marqueeRef.current, { y: edge === 'top' ? '-101%' : '101%' }, 0)
-      .set(marqueeInnerRef.current, { y: edge === 'top' ? '101%' : '-101%' }, 0)
-      .to([marqueeRef.current, marqueeInnerRef.current], { y: '0%' }, 0);
-  };
-
-  const handleMouseLeave = (ev) => {
-    if (!itemRef.current || !marqueeRef.current || !marqueeInnerRef.current) return;
-    const rect = itemRef.current.getBoundingClientRect();
-    const x = ev.clientX - rect.left;
-    const y = ev.clientY - rect.top;
-    const edge = findClosestEdge(x, y, rect.width, rect.height);
-
-    gsap
-      .timeline({ defaults: animationDefaults })
-      .to(marqueeRef.current, { y: edge === 'top' ? '-101%' : '101%' }, 0)
-      .to(marqueeInnerRef.current, { y: edge === 'top' ? '101%' : '-101%' }, 0);
-  };
-
-  return (
-    <div className="flowing-menu__item" ref={itemRef} style={{ borderColor }}>
-      <a
-        className="flowing-menu__item-link"
-        href={link}
-        onMouseEnter={handleMouseEnter}
-        onMouseLeave={handleMouseLeave}
-        style={{ color: textColor, flexDirection: isEven ? 'row-reverse' : 'row' }}
-      >
-        <span className="flowing-menu__title">{text}</span>
-        {subtitle && <span className="flowing-menu__subtitle">{(() => {
-          const mid = Math.ceil(subtitle.length / 2);
-          const breakIdx = subtitle.lastIndexOf(' ', mid);
-          const splitAt = breakIdx > 0 ? breakIdx : mid;
-          return <>{subtitle.slice(0, splitAt)}<br />{subtitle.slice(splitAt).trimStart()}</>;
-        })()}</span>}
-      </a>
-      <div className="flowing-marquee" ref={marqueeRef} style={{ backgroundColor: marqueeBgColor }}>
-        <div className="flowing-marquee__inner-wrap">
-          <div className="flowing-marquee__inner" ref={marqueeInnerRef} aria-hidden="true">
-            {[...Array(repetitions)].map((_, idx) => (
-              <div className="flowing-marquee__part" key={idx} style={{ color: marqueeTextColor }}>
-                <span>{text}</span>
-                {image && (
-                  <div
-                    className="flowing-marquee__img"
-                    style={{ backgroundImage: `url(${image})` }}
-                  />
-                )}
-              </div>
-            ))}
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-export default FlowingMenu;
