+++ src/components/Footer.astro
@@ -39,7 +39,7 @@ const contactInfo = [
       <div class="md:col-span-4 space-y-6">
         <div class="flex items-center gap-4">
           <div class="w-12 h-12 bg-white rounded-full flex items-center justify-center p-2">
-            <img src="/logo.svg" alt="OSIX" class="w-full h-full" />
+            <img src="/full_logo.png" alt="OSIX" class="w-full h-full" />
           </div>
           <h3 class="text-2xl text-white">OSIX</h3>
         </div>
