+++ src/components/Hero.astro
@@ -28,7 +28,7 @@ const t = (k: string) => translations.default[k] || k;
             {t('hero.cta.services')}
             <svg class="btn-arrow-icon" width="16" height="16" viewBox="0 0 16 16" fill="none"><path d="M3 8h10M9 4l4 4-4 4" stroke="currentColor" stroke-width="1.5" stroke-linecap="round" stroke-linejoin="round"/></svg>
           </a>
-          <a href="#contact" class="btn-secondary btn-arrow w-full sm:w-auto min-w-[160px] justify-center">
+          <a href="#contact" class="btn-secondary btn-arrow w-full sm:w-auto min-w-[160px] justify-center" style="border-color: black;">
             {t('hero.cta.contact')}
             <svg class="btn-arrow-icon" width="16" height="16" viewBox="0 0 16 16" fill="none"><path d="M3 8h10M9 4l4 4-4 4" stroke="currentColor" stroke-width="1.5" stroke-linecap="round" stroke-linejoin="round"/></svg>
           </a>
