+++ src/components/ServicesShowcase.astro
@@ -272,12 +272,10 @@ const caseLogos: Record<string, string> = {
     justify-content: center;
     text-align: center;
     flex: 1;
-    gap: 2rem;
     padding: 1rem 0;
   }
 
   .ssc-proof-promo-headline {
-    font-family: 'Orbitron', sans-serif;
     font-weight: 700;
     color: #fff;
     line-height: 1.35;
