+++ src/components/kted/variants/ContactA.astro
@@ -221,7 +221,7 @@ const sectors = [
 </section>
 
 <script>
-  import { initEmailJS, sendEmail } from '../../../utils/emailjs.js';
+  import { initEmailJS, sendEmail } from '../../../utils/emailjs';
   import { trackFormStart, trackFormSubmit } from '../../../utils/analytics';
 
   function initializeKTEDContact() {
