+++ src/components/kted/variants/ContactB.astro
@@ -176,7 +176,7 @@ const sectors = [
 </section>
 
 <script>
-  import { initEmailJS, sendEmail } from '../../../utils/emailjs.js';
+  import { initEmailJS, sendEmail } from '../../../utils/emailjs';
   import { trackFormStart, trackFormSubmit } from '../../../utils/analytics';
 
   function initializeKTEDContact() {
