+++ src/components/kted/variants/ContactC.astro
@@ -199,7 +199,7 @@ const sectors = [
 </section>
 
 <script>
-  import { initEmailJS, sendEmail } from '../../../utils/emailjs.js';
+  import { initEmailJS, sendEmail } from '../../../utils/emailjs';
   import { trackFormStart, trackFormSubmit } from '../../../utils/analytics';
 
   function initializeKTEDContact() {
