+++ src/layouts/Layout.astro
@@ -3,7 +3,6 @@ import { ClientRouter } from 'astro:transitions';
 import Header from '../components/Header.astro';
 import Footer from '../components/Footer.astro';
 import '../styles/global.css';
-import '../styles/globals.css';
 
 interface Props {
   title?: string;
